# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new user and start a session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Authenticate with email and password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return the authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/shortlist/player-ids": {
      "get": {
        "operationId": "ShortlistController_playerIds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortlistPlayerIdsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all shortlisted player ids for the current user",
        "tags": [
          "shortlist"
        ]
      }
    },
    "/shortlist": {
      "get": {
        "operationId": "ShortlistController_list",
        "parameters": [
          {
            "name": "position",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Position"
            }
          },
          {
            "name": "nationality",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Argentina",
              "type": "string"
            }
          },
          {
            "name": "seasonId",
            "required": false,
            "in": "query",
            "description": "Restrict to a specific season id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive name search.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minAge",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxAge",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PlayerListCardDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated list of shortlisted players",
        "tags": [
          "shortlist"
        ]
      },
      "post": {
        "operationId": "ShortlistController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddShortlistDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Player added (or already present)."
          },
          "400": {
            "description": "Shortlist size limit reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Player not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a player to the current user shortlist",
        "tags": [
          "shortlist"
        ]
      }
    },
    "/shortlist/{playerId}": {
      "delete": {
        "operationId": "ShortlistController_remove",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Player removed (idempotent)."
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a player from the current user shortlist",
        "tags": [
          "shortlist"
        ]
      }
    }
  },
  "info": {
    "title": "Scout Panel API",
    "description": "Players, seasons, stats and shortlist endpoints for the scout panel.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT issued by POST /auth/login or POST /auth/register. Paste the token without the \"Bearer\" prefix."
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@scout.local"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "supersecret"
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "example": "Validation failed",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "path": {
            "type": "string",
            "example": "/players"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-10T01:33:00.000Z"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@scout.local"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "admin12345"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 10
          },
          "totalItems": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 5
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "page",
          "pageSize",
          "totalItems",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ]
      },
      "Position": {
        "type": "string",
        "enum": [
          "GK",
          "RB",
          "RWB",
          "CB",
          "LB",
          "LWB",
          "CDM",
          "CM",
          "CAM",
          "RW",
          "RM",
          "LW",
          "LM",
          "ST",
          "CF",
          "SS"
        ]
      },
      "PlayerListCardStatsDto": {
        "type": "object",
        "properties": {
          "matchesPlayed": {
            "type": "number",
            "example": 38
          },
          "goals": {
            "type": "number",
            "example": 14
          },
          "assists": {
            "type": "number",
            "example": 5
          },
          "xGPer90": {
            "type": "number",
            "nullable": true,
            "example": 0.55,
            "description": "Expected Goals per 90 minutes"
          }
        },
        "required": [
          "matchesPlayed",
          "goals",
          "assists",
          "xGPer90"
        ]
      },
      "PlayerListCardCurrentSeasonDto": {
        "type": "object",
        "properties": {
          "club": {
            "type": "string",
            "example": "Boca Juniors"
          },
          "clubLogoUrl": {
            "type": "string",
            "nullable": true
          },
          "league": {
            "type": "string",
            "nullable": true
          },
          "stats": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlayerListCardStatsDto"
              }
            ]
          }
        },
        "required": [
          "club",
          "clubLogoUrl",
          "league",
          "stats"
        ]
      },
      "PlayerListCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Position"
              }
            ]
          },
          "nationality": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "currentSeason": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlayerListCardCurrentSeasonDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "photoUrl",
          "position",
          "nationality",
          "birthDate",
          "currentSeason"
        ]
      },
...
      "ShortlistPlayerIdsDto": {
        "type": "object",
        "properties": {
          "playerIds": {
            "example": [
              "boca-merentiel-001"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "playerIds"
        ]
      },
      "AddShortlistDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "example": "boca-merentiel-001"
          }
        },
        "required": [
          "playerId"
        ]
      }
    }
  }
}